changes requested in comments
